move docs to a new lib
