+++ src/libs/gufo2html/templates/attributes.hbs
