+++ src/libs/gufo2html/templates/general_information.hbs
