+++ src/libs/gufo2html/templates/head.hbs
