+++ src/libs/gufo2html/templates/head_content.hbs
