+++ src/libs/gufo2html/templates/layout.hbs
